requirements.txt
langchain-openai
langgraph
langgraph-supervisor
langchain-mcp-adapters
python-dotenv
jinja2
